18W50Alpha
item save is fix and plyer save kind of fixed still issue on loading player information

+++ Assets/Gina/Scripts/Database/Database.cs
@@ -13,12 +13,13 @@ internal class Database
     internal static Item GetItemByID(string v)
     {
         if (items == null || items.Count <= 0) Refresh();
-        var item = items.Find(x => x.id == v);
+        var item = items.Find(x => x.GetID() == v);
         return item != null ? item : null;
     }
 
     private static List<Item> items = new List<Item>();
-    public static bool saving;
+    public static bool loading_player_data;
+    private static bool saving_player_data;
 
     public static void Refresh()
     {
@@ -51,8 +52,10 @@ internal class Database
 
     internal static void Save(Item value)
     {
+        if (loading_player_data) return;
         new Thread(new ThreadStart(() =>
         {
+            loading_player_data = true;
             if (!string.IsNullOrEmpty(value.file) && File.Exists(value.file))
                 File.Delete(value.file);
             var name = value.Get<string>(Options.name);
@@ -61,55 +64,76 @@ internal class Database
             var savefile = Path.Combine(item_base_folder, $"{name}.json").Replace("\\", "/");
             File.WriteAllText(savefile, json, Encoding.UTF8);
             GameManager.ExecuteAction(Refresh);
-
+            loading_player_data = false;
         })).Start();
     }
     internal static void Save(PlayerData value)
     {
+        if (saving_player_data) return;
         new Thread(new ThreadStart(() =>
         {
+            saving_player_data = true;
+            var statjson = JsonConvert.SerializeObject(value._stats, Formatting.Indented);
+            var invjson = JsonConvert.SerializeObject(value._inventory, Formatting.Indented);
+            var actjson = JsonConvert.SerializeObject(value._actionbar, Formatting.Indented);
+            var charjson = JsonConvert.SerializeObject(value._character, Formatting.Indented);
+            var sb = new StringBuilder();
+            sb.AppendLine(statjson);
+            sb.AppendLine("-");
+            sb.AppendLine(invjson);
+            sb.AppendLine("-");
+            sb.AppendLine(actjson);
+            sb.AppendLine("-");
+            sb.AppendLine(charjson);
+
             var savepath = Path.Combine(Application.streamingAssetsPath, "player_.json").Replace("\\", "/");
             if (!string.IsNullOrEmpty(savepath) && File.Exists(savepath))
-                File.Delete(savepath);
-            var json = JsonConvert.SerializeObject(value, Formatting.Indented);
-            File.WriteAllText(savepath, json, Encoding.UTF8);
+            File.Delete(savepath);
+           
+            File.WriteAllText(savepath, sb.ToString(), Encoding.UTF8);
             Debug.Log("Saved");
             GameManager.ExecuteAction(Refresh);
+            saving_player_data = false;
 
         })).Start();
     }
 
-    internal static void LoadPlayerData(PlayerData into, Action action = null)
+    internal static void LoadPlayerData(PlayerData output, Action action = null)
     {
         new Thread(new ThreadStart(() =>
         {
-            saving = true;
+            loading_player_data = true;
             PlayerData data = null;
             Directory.CreateDirectory(Application.streamingAssetsPath);
             var loadpath = Path.Combine(Application.streamingAssetsPath, "player_.json").Replace("\\", "/");
             if (File.Exists(loadpath))
             {
-                var json = File.ReadAllText(loadpath, Encoding.UTF8);
-                data = JsonConvert.DeserializeObject<PlayerData>(json);
+                data = new PlayerData();
+                var json = File.ReadAllText(loadpath, Encoding.UTF8).Split('-');
+                data._stats = JsonConvert.DeserializeObject<Dictionary<Options, object>>(json[0]);
+                data._inventory = JsonConvert.DeserializeObject<Dictionary<Options, object>[]>(json[1]);
+                data._actionbar = JsonConvert.DeserializeObject<Dictionary<Options, object>[]>(json[2]);
+                data._character = JsonConvert.DeserializeObject<Dictionary<Options, object>[]>(json[3]);
+
             }
             if (data == null)
             {
                 data = new PlayerData();
-                data.stats = PlayerData.DefaultStats;
-                data.inventory = new Item[30];
-                data.actionbar = new Item[12];
-                data.character = new Item[15];
+                data._stats = PlayerData.DefaultStats.data;
+                data._inventory = new Dictionary<Options, object>[30];
+                data._actionbar = new Dictionary<Options, object>[12];
+                data._character = new Dictionary<Options, object>[15];
             }
             else
             {
-                ((PlayerData)(object)into).stats = new Item(data.stats.data);
-                Array.Copy(data.inventory, ((PlayerData)(object)into).inventory, 30);
-                Array.Copy(data.actionbar, ((PlayerData)(object)into).actionbar, 12);
-                Array.Copy(data.character, ((PlayerData)(object)into).character, 15);
+                output._stats = data._stats;
+                output._inventory = data._inventory;
+                output._actionbar = data._actionbar;
+                output._character = data._character;
             }
             if (action != null)
                 GameManager.ExecuteAction(action);
-            saving = false;
+            loading_player_data = false;
         })).Start();
     }
